Add solution for 2022-D13
Parsing is all done manually in this problem. There's a function to
extract the top level items from a string by keeping track of the
brackets passing, and some base cases to compare strings containing
simple arrays. The comparison for generic strings is handled
recursively and is used as input for the sort function in part 2.