Add solution for 2022-D14
In part 1 the map is only made as wide as the outermost rock positions
so that the edges represent spilling into the abyss. To reuse the move
semantics in part 2 the trick is to widen the map to accommodate the
widest possible tree. That way, the spout is blocked before any grain
of sand encounters the boundary.