Add solution for 2022-D15
Part 1 I implemented very quickly, but for a long time I failed to
account for the width of the sensor exclusion zones so I wasn't
counting some points off the edges of the real input.

For part 2 I sketched out some possibilities on paper and quickly
found that we only need to check points just outside the exclusion
zone of each sensor and within the given grid. This was also quick to
implement and I found the correct point in the test input as well as a
unique point in the real input. Took me an embarrassingly long time to
figure out my frequency overflowed right at the end. Oops.