Add solution for 2022-D18
After some tricky second parts of D16 & 17 this puzzle was a
relief. In part 1 I start with every cube exposing 6 faces and
subtract every neighbour that shares a face with it. Initially I tried
to calculate and subtract the internal exposed faces in part 2 from
the total area in part 1. Instead, I put the lava in a padded box and
performed a modified BFS to flood the outside and count whenever a
face is encountered.