Add solution for 2022-D21
The first part was solved with no optimizations, just a recursive
function to evaluate the expressions down to ones with a definite
value.

To speed things up in the second part I determined all the
expressions that were independent of the human value and evaluated
them fully -- a significant reduction for the problem input. This is
how I discovered that half the root expression for my input was
independent of the human value. After doing a quick search for
positive and negative values I do a simple bisection method on the
difference of the root expressions to find where they're equal.