Add solution for 2022-D24
To quickly search the paths later I first store the time evolution of
the storms up to some maximum. Next I perform a BFS indexed by the
position and time, only proposing the move if its not occupied by a
storm after the time step. By keeping track of the time we can
navigate between start and finish with extra searches as required for
part 2.