Add solution for 2022-D25
For a given number of SNAFU digits it's not hard to work out the
maximum & minimum (negative) values it can take. This helps us
establish a correspondence between the SNAFU digits {2,1,0,-,=} and
base-5 digits {4,3,2,1,0}. So, starting with the decimal value we can
shift it to be positive and decompose into base-5 digits before
transforming into the SNAFU symbols. Part 2 is to complete my
outstanding problems.

Merry Christmas :)