Add solution for 2022-D19
It turns out that a BFS works if you severely limit the
candidates. One optimization comes from the fact that we're only
allowed to spend once per turn. If we have more robots of a certain
resource than we can spend in a single turn there's no sense keeping
extra robots because we'll make that up by the time we're ready to
spend on the next turn. Similarly, if we have so many resources that
we couldn't possibly spend them all in the remaining time we can get
rid of that extra stock.